# Conflicts:
#	src/main/java/de/tum/in/www1/artemis/service/UserService.java
#	src/main/java/de/tum/in/www1/artemis/service/connectors/CIUserManagementService.java
#	src/main/java/de/tum/in/www1/artemis/service/connectors/jenkins/JenkinsUserManagementService.java
#	src/main/java/de/tum/in/www1/artemis/service/connectors/jenkins/dto/JenkinsUserDTO.java
#	src/main/webapp/app/overview/course-card.component.ts

+++ src/main/java/de/tum/in/www1/artemis/service/FileService.java
@@ -5,10 +5,7 @@ import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 import java.io.*;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
+import java.nio.file.*;
 import java.time.ZonedDateTime;
 import java.util.Collection;
 import java.util.Map;
@@ -72,7 +69,7 @@ public class FileService implements DisposableBean {
 
     @CacheEvict(value = "files", key = "#path")
     public void resetOnPath(String path) {
-        log.info("Invalidate files cache for " + path);
+        log.info("Invalidate files cache for {}", path);
         // Intentionally blank
     }
 
@@ -345,6 +342,10 @@ public class FileService implements DisposableBean {
             // special case for '.classpath' files which would not be included in the build otherwise
             if (targetFilePath.endsWith("classpath.file")) {
                 targetFilePath = targetFilePath.replace("classpath.file", ".classpath");
+            }
+            // special case for 'dune' files which would not be included in the build otherwise
+            if (targetFilePath.endsWith("dune.file")) {
+                targetFilePath = targetFilePath.replace("dune.file", "dune");
             }
 
             Path copyPath = Paths.get(targetDirectoryPath + targetFilePath);
@@ -666,8 +667,10 @@ public class FileService implements DisposableBean {
     public void scheduleForDeletion(Path path, long delayInMinutes) {
         ScheduledFuture<?> future = executor.schedule(() -> {
             try {
-                log.info("Delete file " + path);
-                Files.delete(path);
+                if (Files.exists(path)) {
+                    log.info("Delete file {}", path);
+                    Files.delete(path);
+                }
                 futures.remove(path);
             }
             catch (IOException e) {
@@ -687,8 +690,10 @@ public class FileService implements DisposableBean {
     public void scheduleForDirectoryDeletion(Path path, long delayInMinutes) {
         ScheduledFuture<?> future = executor.schedule(() -> {
             try {
-                log.info("Delete directory  " + path);
-                FileUtils.deleteDirectory(path.toFile());
+                if (Files.exists(path) && Files.isDirectory(path)) {
+                    log.debug("Delete directory {}", path);
+                    FileUtils.deleteDirectory(path.toFile());
+                }
                 futures.remove(path);
             }
             catch (IOException e) {
@@ -724,9 +729,18 @@ public class FileService implements DisposableBean {
                 Files.createDirectories(uniquePath);
             }
             catch (IOException e) {
-                log.warn("could not create the directories for the path " + uniquePath);
+                log.warn("could not create the directories for the path {}", uniquePath);
             }
         }
         return uniquePath;
+    }
+
+    /**
+     * Removes illegal characters for filenames from the string.
+     * @param string the string with the characters
+     * @return stripped string
+     */
+    public String removeIllegalCharacters(String string) {
+        return string.replaceAll("/[/\\\\?%*:|\"<>]/g", "");
     }
 }
